Разделяет на модули
